AI_Drama_Translator/requirements.txt
PyQt6
opencv-python
numpy
requests
Pillow
pyobjc-framework-Vision
pyobjc-framework-Cocoa
pyobjc-framework-Quartz
